Allow people to use &apos; entity (sinqle quote character: ') in wikitext.
It probably wasn't included originally, as its not part of html4,
but was introduced in XML/XHTML1. Since we do all sorts of weird stuff
with ' characters (italic/bold) its probably good to allow lowly users to
have an easy to remember way of escaping it.

There's also a whole bunch of entities in html5 that we should maybe consider
decoding as well.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@86061 dd0e9695-b195-4be7-bd10-2dea1a65a6b6