Handle old libxml when extracting SVG metadata
Mac OS X 10.5.8 comes with libxml 2.6.16, thus some methods of XMLReader
are not availabe (ex: readInnerXML()).
This patch, throw an error if the above method does not exist (only one
use in our code).

Since metadata can comes as string or an XML fragment, I have split the tests
to take care of the two usages and of the exception.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@89029 dd0e9695-b195-4be7-bd10-2dea1a65a6b6