adjusting width of ArticleFeedback-tooltip container.
- Previously much wider than the actual 5-star-rating-container, which was done as a middle way to have tooltips on one line, but due to the last rating-box not having any space to the right this in-between solution looked inconsistent.
However when making it as wide as the 5-star-rating-container  (11em) some tooltips will wrap three lines: http://i.imgur.com/HHbGW.png

For now made the width equal to that 11em + 1em = 12em where 1em is the margin between the boxes so it's still wider and won't cause tooltips to span 3 lines, and still not underneath other rating boxes either.

This solution is better than what it was but not ideal and especially when i18n comes in, it's simply unacceptable. This will probably be a moot point when the widget is redesigned to take all the recent modifications into account.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@91466 dd0e9695-b195-4be7-bd10-2dea1a65a6b6