(Bug 19725) Do not include suppressed edits in the "View X deleted edits" message, and when doing prefix search of special:undelete.
I'm not 100% sure this is the right thing to do, see the bug for the details. But basically this doesn't include an edit in the count if its text is hidden and its hidden from admins. (Not sure if it should not be included only if everything is hidden). Its also weird to show people different things depending if they have suppress rights, without really indicating that.

Minor db note: This causes the query to no longer use a covering index. I don't think that matters but just thought i'd mention.

p.s. The upload page show deleted edits link is broken right now, (from before) I'll fix in a follow-up.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@91561 dd0e9695-b195-4be7-bd10-2dea1a65a6b6