Set envelope sender to the same as the From: address when PEAR mail is used.  This is so that mail “bounces” to the expected place.
Note that this only works when $wgSMTP is set and (as a result) the PEAR mailer is used.

When $wgSMTP is not set, the envelope sender is set using $wgAdditionalMailParams set to something like “-f from@example.com”.  Someone should create some Envelope sender sanity in how the envelope sender is set.

Suggestion: if $wgAdditionalMailParams is not set, then use $wgEmergencyContact to create “-f $wgEmergencyContact” for php mailer and to set the Return-Path header for PEAR mailer.  I used “From:” for Return-Path here because it seemed most sensible.

See also: http://pear.php.net/bugs/bug.php?id=5017


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@91662 dd0e9695-b195-4be7-bd10-2dea1a65a6b6