* (bug 29804) Fix byte-length-limited fields to check the length of to-be-added char instead of hardcoding a possible length of 3.
Since the Unicode code point of the character being inserted is passed with the keypress event, we can simply turn it into a string and toss it into jquery.byteLength to calculate the length in bytes.

Also fixed a broken test case that had the wrong number of expected chars.

Note that this will probably still not work right in many circumstances -- pastes, anything dealing with selection, and it's entirely possible that individual 'keypress' events might be totally unreliable when dealing with funky input methods.




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@91891 dd0e9695-b195-4be7-bd10-2dea1a65a6b6