Allow blocked sysops to view Special:Unblock
Previously, trying to visit [[Special:Unblock]] while blocked would give
an error, even if the user has the unblockself permission.  I moved the
permission check from execute() to right before the code that does the
actual unblocking.  This should probably be examined closely for
security, since I'm not familiar with this code, although the impact of
a bug would be small.  I tested some simple cases manually and they all
worked as expected:

* Unblocking self works (as before)
* Submitting the form fails unless you're trying to unblock yourself (as
  before)
* GETting the page with any parameters works (previously failed unless
  the target was your own username)

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@92608 dd0e9695-b195-4be7-bd10-2dea1a65a6b6