This extension makes article section edit links clearer.  It:  - moves the links next to the section title without a parser patch  - adds a hover action that highlights the edit link when the mouse is placed on the section  - adds an outline around the section when the mouse hovers over the link
It expands on EditSectionHiliteLink, and requires the new parser hook added in r92506.  Tested with section-spanning DIVs and seems to work in LTR.

Could use some visual cleanup, but the code is here and someone with a more developed sense for design could take it and make it pretty.  For example, that unicode arrow could probably go.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@92919 dd0e9695-b195-4be7-bd10-2dea1a65a6b6