add inline comments to the diff output
This patch rely on various previous revisions made to the CodeRevision
extension. It is mostly experimental, but there is an easy way to
disable it (will need to write a patch for this later so we use a global
variable like $wgEnableInlineComments).

To use it, simply click on a diff line to have a comment box opening.

Stuff that need fixing:
* the comment box is inserted above previous comments to a given line
* anonymous user (or people without the correct rights) will be show the
  comment although the API did not save it in the DB. This is a bug in
  the API code : it does not check the correct right and/or does not
  return an error message if the comment was not inserted.
* Someone without the correct right should not be able to add a comment
* API call returns an HTML formatted comment, might be unneeded sometime
* JavaScript not reviewed for performance
* Not reviewed for security

A patch will come later on to easily disable this feature.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@95435 dd0e9695-b195-4be7-bd10-2dea1a65a6b6