Removed the 'eclipse helper' bit on top of every API module
It's a parctice that dates back to 2006 when the API was first written, and frankly isn't covered by the coding conventions. Same thing with the docblocks, they're all copypasted with some bits changed and don't even make sense if you look at them in the genereated code docs.
I don't feel that any of us depend on this anymore (get a better IDE), so in the inerest of consistancy it's time we said goodbye to it.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@103273 dd0e9695-b195-4be7-bd10-2dea1a65a6b6