Initial commit of improved file page for MobileFrontend
Please test on various devices; may need to be altered or removed for devices with less-capable browsers, but might work fine.
Known good on iOS 5 and Android 2.3 browsers.

Based on previous mockup work.

Reformats the filetoc list and turns the links into toggles for which section to show.
Each section shows by itself, so a wide history table doesn't hose the rest of the page.
Image is fit to the screen, but still using the larger-size screen thumbnail so looks nice and high-res and zoomable.
Disables clickthrough on the image (can still copy the link etc, but tapping won't shove you through to the raw image by accident. You can still tap on the detail links, which still exist)


Similar improvements should be brought to the desktop site, perhaps. :)



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@105768 dd0e9695-b195-4be7-bd10-2dea1a65a6b6