parseChunkDelimiter is now newline agnostic
Under Windows, newline is \r\n. Thus we were passing a chunk delimiter ending
with a remaining \r and the chunk was never recognized. That caused an
exception to be thrown when CodeReview is installed on a Windows server.

rtrim() happilly makes this function agnostic.

added two tests cases to make sure we do not choke anymore on this issue.

Fix r94951



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@106113 dd0e9695-b195-4be7-bd10-2dea1a65a6b6