[JSTesting] several updates
* Sync QUnitTestResources.php with qunit/index.html

* Fix usage of global wg var

* Fix tablesorter.test (to be merged to trunk[1])
-- In the table sorter code it applies certain conditions if the content language is "en", and otherwise to whatever is specified. In qunit/index.html there is no wgContentLanguage set, so it assumed it's not default and used the "dmy" date format. On Special:JavaScriptTest however it refused to use it since wgContentLanguage was set to "en" (as is default in MediaWiki). Changing test suite to declare the language itself, instead of relying on globals.

TODO: Hook into QUnit "setup" and "teardown" for each test() so that globals are reset automatically.
This is actually pretty cool, as for mw.config/QUnit all we need to do is save mw.config.values in a temp val, and set it to a new empty object in "setup" and restore in "teardown". Instant "complete config scope" mock & restore. 

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/branches/JSTesting@107506 dd0e9695-b195-4be7-bd10-2dea1a65a6b6