Reverts MySQL stored procedure support
This is reverting the work done by MaxSem to support stored procedures
and stored function in MySQL. The reasons are:
 - it is not needed yet
 - tests are not functionals
 - alter the stable include/db/Database.php and drop support for ';;'

So please create a branch to work on it and merge it back in trunk
once we have branched 1.19 :-)

I have opened bug 33654 to track this enhancement request.

Reverts r107376, r107994.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@108603 dd0e9695-b195-4be7-bd10-2dea1a65a6b6