Followup r109711
When given a script, it may or may not have the full path. So cannot do a direct array comparison. Which means it can be: $relFile, $maintenance . $relFile, $wikimediaMaintenance . $relFile

Try all locations, and if no matches to the wikiless presume it needs a wiki parameter (or else, it's a script we haven't been told is wikiless yet, tough! add it yourself)

The scripts that have moved for 1.19 can be removed from the phase3/maintenance $wikiless list after 1.19 is sitewide

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@109739 dd0e9695-b195-4be7-bd10-2dea1a65a6b6