diff: align cell content to the top
This is needed when both sides are unbalanced. For example when a very
long paragraph is replaced by a single sentence, the text in the right
cell will be floating in the middle.

Example using ?action=render-ascii:

PRE PATCH:
+----------------+  +----------------+
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  | lot of e       |
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  |                |
+----------------+  +----------------+

POST PATCH:
+----------------+  +----------------+
|eeeeeeeeeeeeeeee|  | lot of e       |
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  |                |
|eeeeeeeeeeeeeeee|  |                |
+----------------+  +----------------+


Part of: bug 33335 - new color scheme and text display for diffs
follow r107127



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@109932 dd0e9695-b195-4be7-bd10-2dea1a65a6b6