lose focus on click bug and adjust search element and styling for iphone
when you type in the search box and click back the text of the input
is clipped on iphone. This seems to be because Safari adds padding to
default styling.

also fix issue where iphone/safari lose focus on click
on clicking on the search area in safari/iphone the search box
loses focus and the user must click again before typing
this appears to be a glitch in Safari
this can be remedied by changing the search input type from text
to search




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@112025 dd0e9695-b195-4be7-bd10-2dea1a65a6b6