do not run javascript on any browser < IE 10
currently the mobile site breaks on various versions of
Internet Explorer. This is due to lack of support of things
such as addEventListener, getElementsByClassName, document.querySelector
and XMLHttpRequest

It is feasible that certain people in low bandwidth countries may be using
legacy browsers to access the mobile site of mediawiki instances. We shouldn't
break the site for these people and at least give them a non-javascript version in
the meantime and fix this as demand requires.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@112122 dd0e9695-b195-4be7-bd10-2dea1a65a6b6