(bug 34736) empty limit on special pages causes navigation issues
On special pages, the older link showed a limit='' query parameter which
made the page show nothing ('' is evalued to a limit of 0 articles).

The cause is r105809 which changed the way we interprets parameters in
wfArrayToCGI() and wfCGIToArray(). Previously, An empty string value would skip
the key, after r105809, we need to use null.




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@112474 dd0e9695-b195-4be7-bd10-2dea1a65a6b6